oruma-backend-fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic[email]==2.4.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.0.1
python-dotenv==1.0.0
redis==5.0.1
websockets==11.0.3
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0